+++ docs/examples/index.js
@@ -1,7 +1,5 @@
 import button from './button.txt';
 import checkbox from './checkbox.txt';
+import Switch from './switch.txt';
 
-export {
-  button,
-  checkbox
-};
+export { button, checkbox, Switch };