+++ docs/readmes/index.js
@@ -1,7 +1,5 @@
 import button from './button.md';
 import checkbox from './checkbox.md';
+import Switch from './switch.md';
 
-export {
-  button,
-  checkbox
-};
+export { button, checkbox, Switch };