+++ src/index.js
@@ -1,5 +1,6 @@
 import Ripple from './ripple/ripple';
 import Button from './button/button';
 import Checkbox from './checkbox/checkbox';
+import Switch from './switch/switch';
 
-export { Button, Checkbox, Ripple };
+export { Button, Checkbox, Ripple, Switch };